+++ Src/Visual Studio/SDK/C#/Eruru.ChatRobotRPC For Shared Project/Informations/ChatRobotGiftInformation.cs
@@ -8,7 +8,7 @@
 		/// <summary>
 		/// 礼物ID
 		/// </summary>
-		public long ID { get; set; }
+		public int ID { get; set; }
 		/// <summary>
 		/// 礼物名称
 		/// </summary>
@@ -16,15 +16,15 @@
 		/// <summary>
 		/// 礼物价值
 		/// </summary>
-		public long Value { get; set; }
+		public int Value { get; set; }
 		/// <summary>
 		/// 礼物获得时间戳
 		/// </summary>
-		public long GetTimeStamp { get; set; }
+		public int GetTimeStamp { get; set; }
 		/// <summary>
 		/// 礼物过期时间戳
 		/// </summary>
-		public long ExpirationTimeStamp { get; set; }
+		public int ExpirationTimeStamp { get; set; }
 
 	}
